+++ jabgui/src/main/java/org/jabref/gui/frame/JabRefFrameViewModel.java
@@ -89,25 +89,40 @@ public class JabRefFrameViewModel implements UiMessageHandler {
 
     void storeLastOpenedFiles(List<Path> filenames, Path focusedDatabase) {
         if (preferences.getWorkspacePreferences().shouldOpenLastEdited()) {
-            // Here we store the names of all current files. If there is no current file, we remove any
-            // previously stored filename.
-            if (filenames.isEmpty()) {
+            List<Path> pathsToSave = filenames;
+            Path focusedToSave = focusedDatabase;
+
+            if (preferences.getInternalPreferences().isMemoryStickMode()) {
+                Path baseDir = Path.of("").toAbsolutePath().normalize();
+                pathsToSave = filenames.stream()
+                                       .map(path -> {
+                                           try {
+                                               return baseDir.relativize(path.toAbsolutePath().normalize());
+                                           } catch (IllegalArgumentException e) {
+                                               return path;
+                                           }
+                                       })
+                                       .toList();
+
+                if (focusedDatabase != null) {
+                    try {
+                        focusedToSave = baseDir.relativize(focusedDatabase.toAbsolutePath().normalize());
+                    } catch (IllegalArgumentException e) {
+                        focusedToSave = focusedDatabase;
+                    }
+                }
+            }
+
+            if (pathsToSave.isEmpty()) {
                 preferences.getLastFilesOpenedPreferences().getLastFilesOpened().clear();
             } else {
-                preferences.getLastFilesOpenedPreferences().setLastFilesOpened(filenames);
-                preferences.getLastFilesOpenedPreferences().setLastFocusedFile(focusedDatabase);
+                preferences.getLastFilesOpenedPreferences().setLastFilesOpened(pathsToSave);
+                preferences.getLastFilesOpenedPreferences().setLastFocusedFile(focusedToSave);
             }
         }
     }
 
-    /**
-     * Quit JabRef
-     *
-     * @return true if the user chose to quit; false otherwise
-     */
     public boolean close() {
-        // Ask if the user really wants to close, if there are still background tasks running
-        // The background tasks may make changes themselves that need saving.
         if (stateManager.getAnyTasksThatWillNotBeRecoveredRunning().getValue()) {
             Optional<ButtonType> shouldClose = dialogService.showBackgroundProgressDialogAndWait(
                     Localization.lang("Please wait..."),
@@ -118,7 +133,6 @@ public class JabRefFrameViewModel implements UiMessageHandler {
             }
         }
 
-        // Read the opened and focused databases before closing them
         List<Path> openedLibraries = tabContainer.getLibraryTabs().stream()
                                                  .map(LibraryTab::getBibDatabaseContext)
                                                  .map(BibDatabaseContext::getDatabasePath)
@@ -129,12 +143,11 @@ public class JabRefFrameViewModel implements UiMessageHandler {
                                         .flatMap(BibDatabaseContext::getDatabasePath)
                                         .orElse(null);
 
-        // Then ask if the user really wants to close, if the library has not been saved since last save.
         if (!tabContainer.closeTabs(tabContainer.getLibraryTabs())) {
             return false;
         }
 
-        storeLastOpenedFiles(openedLibraries, focusedLibraries); // store only if successfully having closed the libraries
+        storeLastOpenedFiles(openedLibraries, focusedLibraries);
 
         ProcessingLibraryDialog processingLibraryDialog = new ProcessingLibraryDialog(dialogService);
         processingLibraryDialog.showAndWait(tabContainer.getLibraryTabs());
@@ -142,12 +155,6 @@ public class JabRefFrameViewModel implements UiMessageHandler {
         return true;
     }
 
-    /**
-     * Handles commands submitted by the command line or by the remote host to be executed in the ui
-     * Needs to run in a certain order. E.g. databases have to be loaded before selecting an entry.
-     *
-     * @param uiCommands to be handled
-     */
     @Override
     public void handleUiCommands(List<UiCommand> uiCommands) {
         LOGGER.debug("Handling UI commands {}", uiCommands);
@@ -156,12 +163,8 @@ public class JabRefFrameViewModel implements UiMessageHandler {
             return;
         }
 
-        assert !uiCommands.isEmpty();
-
-        // Handle blank workspace
         boolean blank = uiCommands.stream().anyMatch(UiCommand.BlankWorkspace.class::isInstance);
 
-        // Handle OpenDatabases
         if (!blank) {
             uiCommands.stream()
                       .filter(UiCommand.OpenLibraries.class::isInstance)
@@ -187,8 +190,6 @@ public class JabRefFrameViewModel implements UiMessageHandler {
                       .findAny().ifPresent(importBibTex -> importBibtexStringAndOpen(importBibTex.bibtex()));
         }
 
-        // Handle jumpToEntry
-        // Needs to go last, because it requires all libraries opened
         uiCommands.stream()
                   .filter(UiCommand.JumpToEntryKey.class::isInstance)
                   .map(UiCommand.JumpToEntryKey.class::cast)
@@ -196,12 +197,10 @@ public class JabRefFrameViewModel implements UiMessageHandler {
                   .filter(Objects::nonNull)
                   .findAny().ifPresent(entryKey -> {
                       LOGGER.debug("Jump to entry {} requested", entryKey);
-                      // tabs must be present and contents async loaded for an entry to be selected
                       waitForLoadingFinished(() -> jumpToEntry(entryKey));
                   });
     }
 
-    /// @deprecated used by the browser extension only
     private void importBibtexStringAndOpen(String importStr) {
         LOGGER.debug("ImportBibtex {} requested", importStr);
         BackgroundTask.wrap(() -> {
@@ -213,7 +212,6 @@ public class JabRefFrameViewModel implements UiMessageHandler {
                       .executeWith(taskExecutor);
     }
 
-    /// @deprecated used by the browser extension only
     private void importFromFileAndOpen(String location) {
         LOGGER.debug("Import file {} requested", location);
         BackgroundTask.wrap(() -> CliImportHelper.importFile(location, preferences, false))
@@ -223,27 +221,13 @@ public class JabRefFrameViewModel implements UiMessageHandler {
     }
 
     private void checkForBibInUpperDir() {
-        // "Open last edited databases" happened before this call
-        // Moreover, there is not any CLI command (especially, not opening any new tab)
-        // Thus, we check if there are any tabs open.
         if (tabContainer.getLibraryTabs().isEmpty()) {
             firstBibFile().ifPresent(firstBibFile -> openDatabaseAction.get().openFile(firstBibFile));
         }
     }
 
-    /// Use case: User starts `JabRef.bat` or `JabRef.exe`. JabRef should open a "close by" bib file.
-    /// By "close by" a `.bib` file in the current folder or one level up of `JabRef.exe`is meant.
-    ///
-    /// Paths:
-    ///   - `...\{example-dir}\JabRef\JabRef.exe` (Windows)
-    ///   - `.../{example-dir}/JabRef/bin/JabRef` (Linux)
-    ///   - `...\{example-dir}\JabRef\runtime\bin\JabRef.bat` (Windows)
-    ///
-    /// In the example, `...\{example-dir}\example.bib` should be found.
-    ///
-    /// We do NOT go up another level (i.e., everything in `...` is not found)
     private Optional<Path> firstBibFile() {
-        Path absolutePath = Path.of(".").toAbsolutePath();
+        Path absolutePath = Path.of("").toAbsolutePath();
         if (OS.LINUX && absolutePath.startsWith("/usr")) {
             return Optional.empty();
         }
@@ -261,14 +245,13 @@ public class JabRefFrameViewModel implements UiMessageHandler {
         ArrayList<Path> dirsToCheck = new ArrayList<>(2);
         dirsToCheck.add(Path.of(""));
         if (isJabRefExe) {
-            dirsToCheck.add(Path.of("../"));       // directory above `JabRef.exe` directory
+            dirsToCheck.add(Path.of("../"));
         } else if (isJabRefBat) {
-            dirsToCheck.add(Path.of("../../../")); // directory above `runtime\bin\JabRef.bat`
+            dirsToCheck.add(Path.of("../../../"));
         } else if (isJabRef) {
-            dirsToCheck.add(Path.of("../..(/"));   // directory above `bin/JabRef` directory
+            dirsToCheck.add(Path.of("../../"));
         }
 
-        // We want to check dirsToCheck only, not all subdirs (due to unnecessary disk i/o)
         try {
             return dirsToCheck.stream()
                               .map(Path::toAbsolutePath)
@@ -276,8 +259,6 @@ public class JabRefFrameViewModel implements UiMessageHandler {
                               .filter(path -> FileUtil.getFileExtension(path).equals(Optional.of("bib")))
                               .findFirst();
         } catch (UncheckedIOException ex) {
-            // Could be access denied exception - when this is started from the application directory
-            // Therefore log level "debug"
             LOGGER.debug("Could not check for existing bib file {}", dirsToCheck, ex);
             return Optional.empty();
         }
@@ -297,66 +278,49 @@ public class JabRefFrameViewModel implements UiMessageHandler {
             return;
         }
 
-        // Remove invalid databases
         List<ParserResult> invalidDatabases = parserResults.stream()
                                                            .filter(ParserResult::isInvalid)
                                                            .toList();
         final List<ParserResult> failed = new ArrayList<>(invalidDatabases);
         parserResults.removeAll(invalidDatabases);
 
-        // Add parserResult to the currently opened tab
         for (ParserResult parserResult : parserResults) {
             addParserResult(parserResult);
         }
 
         for (ParserResult parserResult : failed) {
             String message = Localization.lang("Error opening file '%0'",
-                    parserResult.getPath().map(Path::toString).orElse("(File name unknown)")) + "\n" +
+                    parserResult.getPath().map(path -> path.toString()).orElse("(File name unknown)")) + "\n" +
                     parserResult.getErrorMessage();
             dialogService.showErrorDialogAndWait(Localization.lang("Error opening file"), message);
         }
     }
 
     private void addParserResult(ParserResult parserResult) {
-        LOGGER.trace("Adding the entries to the open tab.");
         LibraryTab libraryTab = tabContainer.getCurrentLibraryTab();
-
         BackgroundTask<ParserResult> task = BackgroundTask.wrap(() -> parserResult);
         ImportCleanup cleanup = ImportCleanup.targeting(libraryTab.getBibDatabaseContext().getMode(), preferences.getFieldPreferences());
         cleanup.doPostCleanup(parserResult.getDatabase().getEntries());
         ImportEntriesDialog dialog = new ImportEntriesDialog(libraryTab.getBibDatabaseContext(), task);
-
         dialog.setTitle(Localization.lang("Import"));
         dialogService.showCustomDialogAndWait(dialog);
     }
 
     private void waitForLoadingFinished(Runnable runnable) {
-        LOGGER.trace("Waiting for all tabs being loaded");
-
         CompletableFuture<Void> future = new CompletableFuture<>();
-
         List<ObservableBooleanValue> loadings = tabContainer.getLibraryTabs().stream()
                                                             .map(LibraryTab::getLoading)
                                                             .collect(Collectors.toList());
 
-        // Create a listener for each observable
-        ChangeListener<Boolean> listener = (observable, _, _) -> {
-            // Instanceof implicitly checks for null value
+        ChangeListener<Boolean> listener = (observable, oldValue, newValue) -> {
             if (observable instanceof ObservableBooleanValue observableBoolean) {
                 loadings.remove(observableBoolean);
             }
-            if (LOGGER.isTraceEnabled()) {
-                LOGGER.trace("Count of loading tabs: {}", loadings.size());
-                LOGGER.trace("Count of loading tabs really true: {}", loadings.stream().filter(ObservableBooleanValue::get).count());
-            }
             for (ObservableBooleanValue obs : loadings) {
                 if (obs.get()) {
-                    // Exit the listener if any of the observables is still true
                     return;
                 }
             }
-            // All observables are false, complete the future
-            LOGGER.trace("Future completed");
             future.complete(null);
         };
 
@@ -364,13 +328,8 @@ public class JabRefFrameViewModel implements UiMessageHandler {
             obs.addListener(listener);
         }
 
-        LOGGER.trace("Fire once");
-        // Due to concurrency, it might be that the observables are already false, so we trigger one evaluation
-        listener.changed(null, null, false);
-        LOGGER.trace("Waiting for state changes...");
-
+        listener.changed(null, false, false);
         future.thenRun(() -> {
-            LOGGER.debug("All tabs loaded. Jumping to entry.");
             for (ObservableBooleanValue obs : loadings) {
                 obs.removeListener(listener);
             }
@@ -379,10 +338,9 @@ public class JabRefFrameViewModel implements UiMessageHandler {
     }
 
     private void jumpToEntry(String entryKey) {
-        // check current library tab first
         LibraryTab currentLibraryTab = tabContainer.getCurrentLibraryTab();
         List<LibraryTab> sortedTabs = tabContainer.getLibraryTabs().stream()
-                                                  .sorted(Comparator.comparing(tab -> tab != currentLibraryTab))
+                                                  .sorted(Comparator.comparing(tab -> tab != currentLibraryTab, Comparator.reverseOrder()))
                                                   .toList();
         for (LibraryTab libraryTab : sortedTabs) {
             Optional<BibEntry> bibEntry = libraryTab.getDatabase()
@@ -391,14 +349,11 @@ public class JabRefFrameViewModel implements UiMessageHandler {
                                                                           .equals(entryKey))
                                                     .findAny();
             if (bibEntry.isPresent()) {
-                LOGGER.debug("Found entry {} in library tab {}", entryKey, libraryTab);
                 libraryTab.clearAndSelect(bibEntry.get());
                 tabContainer.showLibraryTab(libraryTab);
                 break;
             }
         }
-
-        LOGGER.trace("End of loop");
 
         if (stateManager.getSelectedEntries().isEmpty()) {
             dialogService.notify(Localization.lang("Citation key '%0' to select not found in open libraries.", entryKey));